Move structArray test to its own file
