+++ csharp/test/Apache.Arrow.Tests/ArrayBuilderTests.cs
@@ -74,68 +74,6 @@ namespace Apache.Arrow.Tests
         }
 
         [Fact]
-        public void TestStructArray()
-        {
-            // The following can be improved with a Builder class for StructArray.
-            List<Field> fields = new List<Field>();
-            Field.Builder fieldBuilder = new Field.Builder();
-            fields.Add(fieldBuilder.Name("Strings").DataType(StringType.Default).Nullable(true).Build());
-            fieldBuilder = new Field.Builder();
-            fields.Add(fieldBuilder.Name("Ints").DataType(Int32Type.Default).Nullable(true).Build());
-            StructType structType = new StructType(fields);
-
-            StringArray.Builder stringBuilder = new StringArray.Builder();
-            StringArray stringArray = stringBuilder.Append("joe").AppendNull().AppendNull().Append("mark").Build();
-            Int32Array.Builder intBuilder = new Int32Array.Builder();
-            Int32Array intArray = intBuilder.Append(1).Append(2).AppendNull().Append(4).Build();
-            List<Array> arrays = new List<Array>();
-            arrays.Add(stringArray);
-            arrays.Add(intArray);
-
-            ArrowBuffer.BitmapBuilder nullBitmap = new ArrowBuffer.BitmapBuilder();
-            var nullBitmapBuffer = nullBitmap.Append(true).Append(true).Append(false).Append(true).Build();
-            StructArray structs = new StructArray(structType, 4, arrays, nullBitmapBuffer, 1);
-
-            Assert.Equal(4, structs.Length);
-            Assert.Equal(1, structs.NullCount);
-            ArrayData[] childArrays = structs.Data.Children; // Data for StringArray and Int32Array
-            Assert.Equal(2, childArrays.Length);
-            for (int j = 0; j < childArrays.Length; j++)
-            {
-                ArrayData arrayData = childArrays[j];
-                Assert.Null(arrayData.Children);
-                if (j == 0)
-                {
-                    Assert.Equal(ArrowTypeId.String, arrayData.DataType.TypeId);
-                    Array array = new StringArray(arrayData);
-                    StringArray structStringArray = array as StringArray;
-                    Assert.NotNull(structStringArray);
-                    Assert.Equal(structs.Length, structStringArray.Length);
-                    Assert.Equal(stringArray.Length, structStringArray.Length);
-                    Assert.Equal(stringArray.NullCount, structStringArray.NullCount);
-                    for (int i = 0; i < stringArray.Length; i++)
-                    {
-                        Assert.Equal(stringArray.GetString(i), structStringArray.GetString(i));
-                    }
-                }
-                if (j == 1)
-                {
-                    Assert.Equal(ArrowTypeId.Int32, arrayData.DataType.TypeId);
-                    Array array = new Int32Array(arrayData);
-                    Int32Array structIntArray = array as Int32Array;
-                    Assert.NotNull(structIntArray);
-                    Assert.Equal(structs.Length, structIntArray.Length);
-                    Assert.Equal(intArray.Length, structIntArray.Length);
-                    Assert.Equal(intArray.NullCount, structIntArray.NullCount);
-                    for (int i = 0; i < intArray.Length; i++)
-                    {
-                        Assert.Equal(intArray.GetValue(i), structIntArray.GetValue(i));
-                    }
-                }
-            }
-        }
-
-        [Fact]
         public void ListArrayBuilder()
         {
             var listBuilder = new ListArray.Builder(StringType.Default);
